requirements.txt
ai2thor==4.2.0
clip @ git+https://github.com/openai/CLIP.git@e58d49454c92986a1d2a6a48add2333bbfbeaf51
dm-tree==0.1.7
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.3.0/en_core_web_md-3.3.0-py3-none-any.whl
fvcore==0.1.5.post20220414
gensim==4.2.0
h5py==3.6.0
huggingface-hub==0.5.1
hydra-core==1.1.2
imageio==2.18.0
imageio-ffmpeg==0.4.7
ipykernel==6.13.0
ipython==8.2.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
jupyter-client==7.2.2
jupyter-core==4.10.0
jupyter-packaging==0.12.0
jupyter-server==1.17.0
jupyterlab==3.4.0
jupyterlab-pygments==0.2.2
jupyterlab-server==2.13.0
jupyterlab-widgets==1.1.0
matplotlib==3.5.1
matplotlib-inline==0.1.3
moviepy==1.0.3
notebook==6.4.11
notebook-shim==0.1.0
numpy==1.22.3
open3d==0.15.2
opencv-python==4.5.5.64
opencv-python-headless==4.5.5.64
pandas==1.4.2
Pillow==9.1.0
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
sklearn==0.0
spacy-legacy==3.0.9
spacy-loggers==1.0.2
# tf-models-official==2.9.0
# tf-slim==1.1.0
torch==1.9.0+cu111
torchaudio==0.9.0
torchvision==0.10.0+cu111
tornado==6.1
tqdm==4.64.0
transformers==4.18.0
wandb==0.12.15